+++ internal/domain/wallet/dividend.go
@@ -5,7 +5,7 @@ import "github.com/victorskg/my-wallet/internal/domain/stock"
 type Dividend struct {
 	info               stock.Dividend
 	averagePriceOnDate float64
-	quotasOnDate       float64
-	yeldOnCost         float32
+	quotasOnDate       uint32
+	yeldOnCost         float64
 	totalValue         float64
 }
